+++ theme/js/composables/useOwnerName.js
@@ -1,4 +1,4 @@
-import {ref, Ref, ComputedRef, unref} from 'vue';
+import {ref, unref} from 'vue';
 
 /**
  * A resource, dataset, reuse or any other object owned by an organization or a user.
@@ -10,8 +10,8 @@ import {ref, Ref, ComputedRef, unref} from 'vue';
 
 /**
  *
- * @param {Owned|ComputedRef<Owned>} owned - The resource, dataset or reuse owned by an organization or a user.
- * @return {Ref<string>}
+ * @param {Owned|import("vue").ComputedRef<Owned>} owned - The resource, dataset or reuse owned by an organization or a user.
+ * @return {import("vue").Ref<string>}
  */
 export default function useOwnerName(owned) {
   owned = unref(owned);