+++ theme/js/plugins/i18n.js
@@ -4,20 +4,35 @@
 import config from "../config";
 import { createI18n } from "vue-i18n";
 
-//TODO : use a glob when supported in Parcel 2 : https://github.com/parcel-bundler/parcel/issues/4683
-import en from "../locales/en.json";
-import es from "../locales/es.json";
-import fr from "../locales/fr.json"
+import messages from '@intlify/unplugin-vue-i18n/messages';
+import { getRegisteredTranslations } from "@etalab/udata-front-plugins-helper";
 
-
-export default createI18n({
+const i18n = createI18n({
   legacy: false,
   globalInjection: true,
   locale: config.lang,
-  messages: {
-    en,
-    es,
-    fr
-  },
+  messages,
   formatFallbackMessages: true
 });
+
+let loadedModules = {};
+
+/**
+ * Reload translations from plugins if they aren't already loaded
+ */
+export function reloadLocale() {
+  const translations = getRegisteredTranslations();
+  let messages = {};
+  for (let translation of translations) {
+    if(translation.module) {
+      if (loadedModules[translation.module]) {
+        continue;
+      }
+      loadedModules[translation.module] = true;
+    }
+    messages = {...messages, ...translation.messages[config.lang]};
+  }
+  i18n.global.mergeLocaleMessage(config.lang, messages);
+}
+
+export default i18n;