+++ CHANGELOG.md
@@ -5,7 +5,12 @@
 > **Warning**
 > Refactor of breadcrumb macro to be easier to use. `toolbar_class`, `breadcrum_class`, `toolbar_container` and `breadcrumb_bar` options are now removed.
 
-- Nothing yet
+> **Note** <br>
+> This changes the build system from Parcel to Vite.
+> This also adds a way for udata plugin to register their own vue components
+> and to display them in places chosen by the current theme
+
+- Add resource exploration preview [#169](https://github.com/etalab/udata-front/pull/169)
 
 ## 3.0.1 (2022-12-15)
 
@@ -15,6 +20,7 @@
 - Fix margins for dataset cards [#176](https://github.com/etalab/udata-front/pull/176)
 - Fix text for datasets button in org page [#172](https://github.com/etalab/udata-front/pull/172)
 - Add Portuguese translations [#167](https://github.com/etalab/udata-front/pull/167)
+- Add email address to shared data on oauth authorize page [#175](https://github.com/etalab/udata-front/pull/175)
 
 ## 3.0.0 (2022-11-14)