+++ CHANGELOG.md
@@ -7,7 +7,7 @@
 > This also adds a way for udata plugin to register their own vue components
 > and to display them in places chosen by the current theme
 
-- Add resource exploration preview [#169](https://github.com/etalab/udata-front/pull/169)[#180](https://github.com/etalab/udata-front/pull/180)
+- Add resource exploration preview [#169](https://github.com/etalab/udata-front/pull/169)[#180](https://github.com/etalab/udata-front/pull/180)[#183](https://github.com/etalab/udata-front/pull/183)
 - Fix RGAA criterion 9.2 [#178](https://github.com/etalab/udata-front/pull/178)
 - Add i18n on size suffix [#184](https://github.com/etalab/udata-front/pull/184)