+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Add resource exploration preview [#169](https://github.com/etalab/udata-front/pull/169)[#180](https://github.com/etalab/udata-front/pull/180)[#183](https://github.com/etalab/udata-front/pull/183)
 - Fix RGAA criterion 9.2 [#178](https://github.com/etalab/udata-front/pull/178)
 - Add i18n on size suffix [#184](https://github.com/etalab/udata-front/pull/184)
+- Add theme view for mail change [#192](https://github.com/etalab/udata-front/pull/192)
 
 ## 3.0.1 (2022-12-15)